getComputerNames moved to ComputerSet and test accordingly
