+++ core/src/main/java/hudson/cli/ConnectNodeCommand.java
@@ -25,15 +25,13 @@ package hudson.cli;
 
 import hudson.AbortException;
 import hudson.Extension;
-import hudson.Util;
 import hudson.model.Computer;
-import org.acegisecurity.AccessDeniedException;
+import hudson.model.ComputerSet;
 import hudson.util.EditDistance;
 import jenkins.model.Jenkins;
 import org.kohsuke.args4j.Argument;
 import org.kohsuke.args4j.Option;
 
-import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.logging.Logger;
@@ -76,7 +74,7 @@ public class ConnectNodeCommand extends CLICommand {
 
                 if(computer == null) {
                     if(names == null) {
-                        names = Util.getComputerNames();
+                        names = ComputerSet.getComputerNames();
                     }
                     String adv = EditDistance.findNearest(node_s, names);
                     throw new IllegalArgumentException(adv == null ?
