+++ core/src/main/java/hudson/cli/OnlineNodeCommand.java
@@ -26,18 +26,15 @@ package hudson.cli;
 
 import hudson.AbortException;
 import hudson.Extension;
-import hudson.Util;
 import hudson.model.Computer;
+import hudson.model.ComputerSet;
 import hudson.util.EditDistance;
 import jenkins.model.Jenkins;
 
-import org.acegisecurity.AccessDeniedException;
 import org.kohsuke.args4j.Argument;
 
-import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
-import java.util.logging.Logger;
 
 /**
  * @author pjanouse
@@ -68,7 +65,7 @@ public class OnlineNodeCommand extends CLICommand {
                 computer = jenkins.getComputer(node_s);
                 if (computer == null) {
                     if (names == null) {
-                        names = Util.getComputerNames();
+                        names = ComputerSet.getComputerNames();
                     }
                     String adv = EditDistance.findNearest(node_s, names);
                     throw new IllegalArgumentException(adv == null ?
