--- test/src/test/java/hudson/UtilTest.java
@@ -1,55 +1,0 @@
-/*
- * The MIT License
- *
- * Copyright 2016 Red Hat, Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
-package hudson;
-
-import org.junit.Rule;
-import org.junit.Test;
-import org.jvnet.hudson.test.JenkinsRule;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.contains;
-import static org.hamcrest.Matchers.containsInAnyOrder;
-import static org.hamcrest.Matchers.empty;
-import static org.hamcrest.Matchers.hasSize;
-import static org.hamcrest.Matchers.is;
-
-/**
- * @author pjanouse
- */
-public class UtilTest {
-    @Rule
-    public final JenkinsRule j = new JenkinsRule();
-
-    @Test
-    public void testCetComputerNames() throws Exception {
-        assertThat(Util.getComputerNames(), is(empty()));
-        j.createSlave("aNode", "", null);
-        assertThat(Util.getComputerNames(), hasSize(1));
-        assertThat(Util.getComputerNames(), contains("aNode"));
-        j.createSlave("anAnotherNode", "", null);
-        assertThat(Util.getComputerNames(), hasSize(2));
-        assertThat(Util.getComputerNames(), containsInAnyOrder("aNode", "anAnotherNode"));
-    }
-}
